requirements.txt
argparse==1.2.1
requests==1.2.2
wsgiref==0.1.2
xively-python==0.1.0-rc0
